src/whisper-v3-large/requirements.txt
openai-whisper==20231117
transformers==4.39.3
fastapi==0.111.0
python-multipart==0.0.9
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
uvicorn==0.30.1
pydub==0.25.1
whisperx==3.1.3
torchaudio==2.3.0
torchvision==0.18.0
accelerate @ git+https://github.com/huggingface/accelerate.git@v0.26.1
optimum @ git+https://github.com/huggingface/optimum.git@v1.16.1
GPUtil==1.4.0
numpy==1.26.4
nltk==3.8.1
yt-dlp